Added build_installer.bat
The script will build the solution using msbuild and then create
the msi installer using wix.

I have updated the Source locations in the .wxs file to avoid
having to copy files around which was confusing and easy to mess up.

Currently hardcoded to use the Release versions of the files and
using my tool locations by default.

But note: Does not handle/build the .jar file atm.